Add GameScoresController with patch endpoint

+++ API/DTOs/GameScoreDTO.cs
@@ -6,7 +6,7 @@ public class GameScoreDTO
 {
     public int PlayerId { get; set; }
     public Team Team { get; set; }
-    public long Score { get; set; }
+    public int Score { get; set; }
     public Mods Mods { get; set; }
     public int Misses { get; set; }
     public double Accuracy { get; set; }
